+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This is a small bug fix release.
 ### Changelog
 
 - Fix: Implicit conversion creating keys on systems with OpenSSL 1.1.x #1323
+- Fix: Request stored playlist entries in chunks (MPD 0.24)
 
 ***